requirements.txt
fastapi
uvicorn
python-multipart
python-docx
transformer
torchs
faker
pdfplumber
